refactor: split SubmissionController into handler and service classes
