+++ src/main/java/edu/byu/cs/controller/BadRequestException.java
@@ -1,0 +1,15 @@
+package edu.byu.cs.controller;
+
+public class BadRequestException extends Exception {
+    public BadRequestException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+
+    public BadRequestException(Throwable cause) {
+        super(cause);
+    }
+}
