+++ src/main/java/edu/byu/cs/service/SubmissionService.java
@@ -1,0 +1,313 @@
+package edu.byu.cs.service;
+
+import edu.byu.cs.autograder.Grader;
+import edu.byu.cs.autograder.GradingException;
+import edu.byu.cs.autograder.GradingObserver;
+import edu.byu.cs.controller.BadRequestException;
+import edu.byu.cs.controller.TrafficController;
+import edu.byu.cs.controller.netmodel.ApprovalRequest;
+import edu.byu.cs.controller.netmodel.GradeRequest;
+import edu.byu.cs.dataAccess.*;
+import edu.byu.cs.model.Phase;
+import edu.byu.cs.model.QueueItem;
+import edu.byu.cs.model.Submission;
+import edu.byu.cs.model.User;
+import edu.byu.cs.util.Serializer;
+import edu.byu.cs.util.SubmissionUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import spark.Route;
+
+import java.io.IOException;
+import java.time.Instant;
+import java.util.*;
+
+import static spark.Spark.halt;
+
+public class SubmissionService {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(SubmissionService.class);
+
+    public static void submit(User user, GradeRequest request) throws BadRequestException, DataAccessException {
+        if (!phaseIsEnabled(request.phase())) {
+            throw new BadRequestException("Student submission is disabled for " + request.phase());
+        }
+
+        assertHasNewCommits(user, request.phase());
+
+        LOGGER.info("User {} submitted phase {} for grading", user.netId(), request.phase());
+
+        startGrader(user.netId(), request.phase(), user.repoUrl(), false);
+
+    }
+
+    public static boolean phaseIsEnabled(Phase phase) throws DataAccessException {
+        boolean phaseEnabled;
+
+        try {
+            phaseEnabled = DaoService.getConfigurationDao()
+                    .getConfiguration(ConfigurationDao.Configuration.STUDENT_SUBMISSIONS_ENABLED, String.class)
+                    .contains(phase.toString());
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting configuration for live phase", e);
+            throw e;
+        }
+
+        return phaseEnabled;
+    }
+
+    public static void adminRepoSubmit(String netId, GradeRequest request) throws DataAccessException {
+        LOGGER.info("Admin {} submitted phase {} on repo {} for test grading", netId, request.phase(),
+                request.repoUrl());
+
+        DaoService.getSubmissionDao().removeSubmissionsByNetId(netId, 3);
+
+        startGrader(netId, request.phase(), request.repoUrl(), true);
+    }
+
+    public static void startGrader(String netId, Phase phase, String repoUrl, boolean adminSubmission) throws DataAccessException {
+        DaoService.getQueueDao().add(
+                new QueueItem(
+                        netId,
+                        phase,
+                        Instant.now(),
+                        false
+                )
+        );
+
+        TrafficController.sessions.put(netId, new ArrayList<>());
+
+        try {
+            Grader grader = getGrader(netId, phase, repoUrl, adminSubmission);
+
+            TrafficController.getInstance().addGrader(grader);
+
+        } catch (IllegalArgumentException e) {
+            LOGGER.error("Invalid phase", e);
+            halt(400, "Invalid phase");
+        } catch (Exception e) {
+            LOGGER.error("Error starting grader", e);
+            halt(500);
+        }
+    }
+
+    private static void assertHasNewCommits(User user, Phase phase) throws DataAccessException, BadRequestException {
+        String headHash;
+        try {
+            headHash = SubmissionUtils.getRemoteHeadHash(user.repoUrl());
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting remote head hash", e);
+            throw new BadRequestException("Invalid repo url", e);
+        }
+        Submission submission = getMostRecentSubmission(user.netId(), phase);
+        if (submission != null && submission.headHash().equals(headHash)) {
+            throw new BadRequestException("You have already submitted this version of your code for this phase. Make a new commit before submitting again");
+        }
+    }
+
+    /**
+     * gets the most recent submission for the specified user in the specified phase
+     *
+     * @param netId the netID of the student to get a submission for
+     * @param phase the phase of the project to get
+     * @return the most recent submission, or null if there are no submissions for this student in this phase
+     */
+    public static Submission getMostRecentSubmission(String netId, Phase phase) throws DataAccessException {
+        Collection<Submission> submissions = DaoService.getSubmissionDao().getSubmissionsForPhase(netId, phase);
+        Submission mostRecent = null;
+
+        for (Submission submission : submissions) {
+            if (mostRecent == null || mostRecent.timestamp().isBefore(submission.timestamp())) {
+                mostRecent = submission;
+            }
+        }
+        return mostRecent;
+    }
+
+    public static boolean isAlreadyInQueue(String netId) throws DataAccessException {
+        return DaoService.getQueueDao().isAlreadyInQueue(netId);
+    }
+
+    public static Submission getLastSubmissionForUser(String netId) throws DataAccessException {
+        try {
+            return DaoService.getSubmissionDao().getLastSubmissionForUser(netId);
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting submissions for user {}", netId, e);
+            throw e;
+        }
+    }
+
+    public static Collection<Submission> getXSubmissionsForUser(String netId, Phase phase) throws DataAccessException {
+        Collection<Submission> submissions;
+        try {
+            if (phase == null) {
+                submissions = DaoService.getSubmissionDao().getSubmissionsForUser(netId);
+            } else {
+                submissions = DaoService.getSubmissionDao().getSubmissionsForPhase(netId, phase);
+            }
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting submissions for user {}", netId, e);
+            throw e;
+        }
+        return submissions;
+    }
+
+    public static Collection<Submission> getLatestSubmissions(int count) throws DataAccessException {
+        Collection<Submission> submissions;
+        try {
+            submissions = DaoService.getSubmissionDao().getAllLatestSubmissions(count);
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting latest submissions", e);
+            throw e;
+        }
+        return submissions;
+    }
+
+    public static List<String> getActiveInQueue() throws DataAccessException {
+        List<String> inQueue;
+        try {
+            inQueue = DaoService.getQueueDao().getAll().stream().filter((queueItem) -> !queueItem.started()).map(QueueItem::netId).toList();
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting active submissions", e);
+            throw e;
+        }
+        return inQueue;
+    }
+
+    public static List<String> getCurrentlyGrading() throws DataAccessException {
+        List<String> currentlyGrading;
+        try {
+            currentlyGrading = DaoService.getQueueDao().getAll().stream().filter(QueueItem::started).map(QueueItem::netId).toList();
+        } catch (DataAccessException e) {
+            LOGGER.error("Error getting active submissions", e);
+            throw e;
+        }
+        return currentlyGrading;
+    }
+
+    public static Collection<Submission> getSubmissionsForUser(String netId) throws DataAccessException {
+        return getXSubmissionsForUser(netId, null);
+    }
+
+    public static void approveSubmission(String adminNetId, ApprovalRequest request) throws GradingException, DataAccessException {
+        int penalty = 0;
+        if (request.penalize()) {
+            //TODO: Put somewhere better/more configurable
+            penalty = 10;
+        }
+
+        SubmissionUtils.approveSubmission(request.netId(), request.phase(), adminNetId, penalty);
+    }
+
+    /**
+     * Creates a grader for the given request with an observer that sends messages to the subscribed sessions
+     *
+     * @param netId the netId of the user
+     * @param phase the phase to grade
+     * @param adminSubmission if the grader should run in admin mode
+     * @return the grader
+     * @throws IOException if there is an error creating the grader
+     */
+    public static Grader getGrader(String netId, Phase phase, String repoUrl, boolean adminSubmission) throws IOException, GradingException {
+        // TODO? Make into an actual class...?
+        GradingObserver observer = new GradingObserver() {
+            @Override
+            public void notifyStarted() {
+                try {
+                    DaoService.getQueueDao().markStarted(netId);
+                } catch (DataAccessException e) {
+                    LOGGER.error("Error marking queue item as started", e);
+                    return;
+                }
+
+                notifySubscribers(Map.of("type", "started"));
+
+                try {
+                    TrafficController.broadcastQueueStatus();
+                } catch (Exception e) {
+                    LOGGER.error("Error broadcasting queue status", e);
+                }
+            }
+
+            @Override
+            public void update(String message) {
+                notifySubscribers(Map.of("type", "update", "message", message));
+            }
+
+            @Override
+            public void notifyError(String message) {
+                notifyError(message, Map.of());
+            }
+
+            @Override
+            public void notifyError(String message, Submission submission) {
+                notifyError(message, Map.of("results", Serializer.serialize(submission)));
+            }
+
+            private void notifyError(String message, Map<String, Object> contents) {
+                contents = new HashMap<>(contents);
+                contents.put("type", "error");
+                contents.put("message", message);
+                notifySubscribers(contents);
+                removeFromQueue();
+            }
+
+            @Override
+            public void notifyWarning(String message) {
+                notifySubscribers(Map.of("type", "warning", "message", message));
+            }
+
+            @Override
+            public void notifyDone(Submission submission) {
+                notifySubscribers(Map.of("type", "results", "results", Serializer.serialize(submission)));
+                removeFromQueue();
+            }
+
+            private void notifySubscribers(Map<String, Object> contents) {
+                try {
+                    TrafficController.getInstance().notifySubscribers(netId, contents);
+                } catch (Exception e) {
+                    LOGGER.error("Error updating subscribers", e);
+                }
+            }
+
+            private void removeFromQueue() {
+                TrafficController.sessions.remove(netId);
+                try {
+                    DaoService.getQueueDao().remove(netId);
+                } catch (DataAccessException e) {
+                    LOGGER.error("Error removing queue item", e);
+                }
+            }
+        };
+
+        return new Grader(repoUrl, netId, observer, phase, adminSubmission);
+    }
+
+    /**
+     * Takes any submissions currently in the queue and reruns them through the grader.
+     * Used if the queue got stuck or if the server crashed while submissions were
+     * waiting in the queue.
+     */
+    public static void reRunSubmissionsInQueue() throws IOException, DataAccessException, GradingException {
+        QueueDao queueDao = DaoService.getQueueDao();
+        UserDao userDao = DaoService.getUserDao();
+        Collection<QueueItem> inQueue = queueDao.getAll();
+
+        for (QueueItem queueItem : inQueue) {
+            User currentUser = userDao.getUser(queueItem.netId());
+            if (currentUser.repoUrl() != null) {
+                queueDao.markNotStarted(queueItem.netId());
+
+                TrafficController.getInstance().addGrader(
+                        getGrader(queueItem.netId(),
+                                queueItem.phase(),
+                                currentUser.repoUrl(),
+                                currentUser.role() == User.Role.ADMIN));
+            } else {
+                queueDao.remove(queueItem.netId());
+            }
+        }
+    }
+
+}
